Visita con Notificaciones + Seguridad + Usuarios {Nueva forma de crear casos}
Lo mismo que en espera. Ahora los casos conservan el ID cuando se aceptan, rechachan o editan, para poder rastrearlos fácilmente en las notificaciones.

+++ Servidor/controllers/casoVisitaController.js
@@ -2,12 +2,18 @@ const casoVisitaModel = require('../models/casoVisitaModel')
 const casoActivoModel = require('../models/casoActivoModel')
 const casoRechazadoModel = require('../models/casoRechazadoModel')
 const usuarioModel = require('../models/usuarioModel')
+const auth = require('./authController')
 const uuidv4 = require('uuid/v4');
 const crypto = require('crypto');
 const mongoose = require('mongoose')
 const path = require('path');
 
 function getCasosVisita(req, res) {
+  /*if(req.query.token == "undefined" || !auth.autentificarAccion(req.query.token)){
+    res.status(100)
+    res.json({ error: true , casos: []})
+    return
+  }*/
   casoVisitaModel.find()
     .exec((err, casos) => {
       if (err) {
@@ -21,12 +27,25 @@ function getCasosVisita(req, res) {
 
 function createCasoVisita(req, res) {
   //Toma el caso del body (que viene en form data)
-  console.log(req.body)
+  let usuario = JSON.parse(req.body.usuario);
+
+  if(usuario.token===undefined){
+    res.status(500)
+    res.send({ error: true , type: 0})
+    return
+  }
+
+  if(!auth.autentificarAccion(usuario.token)){
+    res.status(500)
+    res.send({ error: true , type: 1})
+    return
+  }
+
   let info = JSON.parse(req.body.caso);
   info["files"] = []
   //Crea caso
   let newCaso = new casoVisitaModel(info)
-  let notificacion = { autor: "kevin", _id: uuidv4(), fecha: new Date(), location: "visita", action: "create", caseId: newCaso._id }
+  let notificacion = { autor: usuario.usuario, _id: uuidv4(), fecha: new Date(), location: "visita", action: "create", caso: newCaso._id }
   newCaso.save((err, resp) => {
     if (err) {
       res.status(500)
@@ -89,17 +108,31 @@ function createCasoVisita(req, res) {
 
 function editCasoVisita(req, res) {
   //Toma el caso del body (que viene en form data)
+  let usuario = JSON.parse(req.body.usuario);
+
+  if(usuario.token===undefined){
+    res.status(500)
+    res.send({ error: true , type: 0})
+    return
+  }
+  if(!auth.autentificarAccion(usuario.token)){
+    res.status(500)
+    res.send({ error: true , type: 1})
+    return
+  }
   let info = JSON.parse(req.body.caso);
-  let notificacion = { autor: "kevin", _id: uuidv4(), fecha: new Date(), location: "visita", action: "update", caseId: info._id }
-  casoVisitaModel.findByIdAndUpdate({ _id: new mongoose.Types.ObjectId(info._id) }, { $set: info})
+  let notificacion = { autor: usuario.usuario, _id: uuidv4(), fecha: new Date(), location: "visita", action: "update", caso: {} }
+  casoVisitaModel.findByIdAndUpdate({ _id: new mongoose.Types.ObjectId(info._id) }, { $set: info}, {new:true})
     .exec((err, caso) => {
       if (err) {
         res.status(500)
         res.send({ error: false })
       }
       else {
-        info["files"] = []
+        caso["files"] = []
         //Recorre req.files en caso de que se haya subido algo
+        notificacion.caso=caso._id
+
         var files = [];
         var archivos = [];
         if (req.files != undefined) {
@@ -109,7 +142,7 @@ function editCasoVisita(req, res) {
           });
         }else{
           res.status(200)
-          res.send({error:false, caso:info})
+          res.send({error:false, caso:caso})
         }
         var i = 0;
         while (i < files.length) {
@@ -128,7 +161,7 @@ function editCasoVisita(req, res) {
               if(caso.files.length>0){
                 archivos = [caso.files, archivos]
               }
-              casoVisitaModel.updateOne({ _id: new mongoose.Types.ObjectId(info._id) }, { $set: { "files": archivos } })
+              casoVisitaModel.updateOne({ _id: new mongoose.Types.ObjectId(caso._id) }, { $set: { "files": archivos } })
                 .exec((err, casod) => {
                   if (err) {
                     res.status(500)
@@ -136,7 +169,7 @@ function editCasoVisita(req, res) {
                   }
                   else {
                     res.status(200)
-                    res.send({ error: false, caso: { ...info, ingreso: new Date(info.ingreso), files: archivos } })
+                    res.send({ error: false, caso: { ...caso, ingreso: new Date(caso.ingreso), files: archivos } })
                   }
                 })
             }
@@ -156,7 +189,19 @@ function editCasoVisita(req, res) {
 }
 
 function acceptCasoVisita(req, res) {
-  console.log(req.body.caso)
+  let usuario = req.body.usuario;
+
+  if(usuario.token===undefined){
+    res.status(500)
+    res.send({ error: true , type: 0})
+    return
+  }
+
+  if(!auth.autentificarAccion(usuario.token)){
+    res.status(500)
+    res.send({ error: true , type: 1})
+    return
+  }
   casoVisitaModel.deleteOne({ _id: new mongoose.Types.ObjectId(req.params.id) })
     .exec((err, caso) => {
       if (err) {
@@ -164,11 +209,11 @@ function acceptCasoVisita(req, res) {
         res.send(`Ocurrió un error 💩 ${err}`)
       }
       let newCaso = new casoActivoModel({
-        cedula: req.body.caso.cedula, apellidos: req.body.caso.apellidos, alternativas: req.body.caso.alternativas,
+        _id:new mongoose.Types.ObjectId(req.params.id),cedula: req.body.caso.cedula, apellidos: req.body.caso.apellidos, alternativas: req.body.caso.alternativas,
         nombre: req.body.caso.nombre, domicilio: req.body.caso.domicilio, telefono: req.body.caso.telefono, nacimiento: req.body.caso.nacimiento,
         ingreso: req.body.caso.ingreso, sede: req.body.caso.sede, señas: req.body.caso.señas, riesgo: req.body.caso.riesgo, notas: req.body.caso.notas
       })
-      let notificacion = { autor: "kevin", _id: uuidv4(), fecha: new Date(), location: "visita", action: "accepted", caseId: newCaso._id }
+      let notificacion = { autor: "kevin", _id: uuidv4(), fecha: new Date(), location: "visita", action: "accepted", caso: newCaso._id }
       newCaso.save((err, resp) => {
         if (err) {
           res.status(500)
@@ -184,6 +229,19 @@ function acceptCasoVisita(req, res) {
 }
 
 function rejectCasoVisita(req, res) {
+  let usuario = req.body.usuario;
+
+  if(usuario.token===undefined){
+    res.status(500)
+    res.send({ error: true , type: 0})
+    return
+  }
+
+  if(!auth.autentificarAccion(usuario.token)){
+    res.status(500)
+    res.send({ error: true , type: 1})
+    return
+  }
   casoVisitaModel.deleteOne({ _id: new mongoose.Types.ObjectId(req.params.id) })
     .exec((err, caso) => {
       if (err) {
@@ -191,11 +249,11 @@ function rejectCasoVisita(req, res) {
         res.send(`Ocurrió un error 💩 ${err}`)
       }
       let newCaso = new casoRechazadoModel({
-        cedula: req.body.caso.cedula, apellidos: req.body.caso.apellidos,
+        _id:new mongoose.Types.ObjectId(req.params.id), cedula: req.body.caso.cedula, apellidos: req.body.caso.apellidos,
         nombre: req.body.caso.nombre, domicilio: req.body.caso.domicilio, señas: req.body.caso.señas, telefono: req.body.caso.telefono,
         sede: req.body.caso.sede, notas: req.body.caso.notas,  nacimiento: req.body.caso.nacimiento, ingreso: req.body.caso.ingreso, 
       })
-      let notificacion = { autor: "kevin", _id: uuidv4(), fecha: new Date(), location: "visita", action: "rejected", caseId: newCaso._id }
+      let notificacion = { autor: "kevin", _id: uuidv4(), fecha: new Date(), location: "visita", action: "rejected", caso: newCaso._id }
       newCaso.save((err, resp) => {
         if (err) {
           res.status(500)
